move criminal and user name getting methods to discordUtil
